+++ source/about.ts
@@ -112,7 +112,7 @@ export default (app: Courselore): void => {
                           onload="${javascript`
                             (this.tooltip ??= tippy(this)).setProps({
                               touch: false,
-                              content: "Sign up on a Courselore installation managed by the developers of Courselore. Enjoy the initial period of free hosting!",
+                              content: "Sign up on a Courselore installation managed by the developers of Courselore.",
                             });
                           `}"
                         >
@@ -126,7 +126,7 @@ export default (app: Courselore): void => {
                           onload="${javascript`
                             (this.tooltip ??= tippy(this)).setProps({
                               touch: false,
-                              content: "Sign in on a Courselore installation managed by the developers of Courselore. Enjoy the initial period of free hosting!",
+                              content: "Sign in on a Courselore installation managed by the developers of Courselore.",
                             });
                           `}"
                         >
@@ -820,7 +820,7 @@ export default (app: Courselore): void => {
                     onload="${javascript`
                       (this.tooltip ??= tippy(this)).setProps({
                         touch: false,
-                        content: "Sign up on a Courselore installation managed by the developers of Courselore. Enjoy the initial period of free hosting!",
+                        content: "Sign up on a Courselore installation managed by the developers of Courselore.",
                       });
                     `}"
                   >
@@ -833,7 +833,7 @@ export default (app: Courselore): void => {
                     onload="${javascript`
                       (this.tooltip ??= tippy(this)).setProps({
                         touch: false,
-                        content: "Sign in on a Courselore installation managed by the developers of Courselore. Enjoy the initial period of free hosting!",
+                        content: "Sign in on a Courselore installation managed by the developers of Courselore.",
                       });
                     `}"
                   >